Moved logic into src/logic, began jobBlock object
