+++ src/logic/jobBlock.js
@@ -1,0 +1,14 @@
+/*
+    jobBlock.js
+    Corbin Weiss
+    2024-4-23
+    Define a job block for the MLFQ graph
+*/
+
+class JobBlock {
+    static start  = 0;  // start time
+    static length = 0;  // length of block
+    static jobID  = 0;  // job ID
+    static JobName
+    static color  = 0;  // color 
+}
